+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Extract hardcoded labels. Fixes UICHKIN-14.
 * Use the new check-in-by-barcode API. Fixes UICHKIN-59.
 * Handle check in for items that are In transit. Fixes UICHKIN-44.
+* Display destination service point for in transit items. Fixes UICHKIN-40.
 
 ## [1.3.0](https://github.com/folio-org/ui-checkin/tree/v1.3.0) (2018-10-04)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.2.0...v1.3.0)